Move query operators into a selector folder, switch from arrow functions to allow for .call()/.apply(), add update modifiers.
