+++ operators/selectors/Element.js
@@ -1,18 +1,18 @@
 /*!
  * A MongoDB inspired ES6 Map() QL. - Copyright (c) 2017 Louis T. (https://lou.ist/)
- * Licensed under MIT license https://raw.githubusercontent.com/LouisT/MapQL/master/LICENSE
+ * Licensed under the MIT license https://raw.githubusercontent.com/LouisT/MapQL/master/LICENSE
  */
 'use strict';
-const Helpers = require('../lib/Helpers');
+const Helpers = require('../../lib/Helpers');
 
 module.exports = {
     '$exists': {
-        chain: (key, val) => {
+        chain: function (key, val) {
             let isbool = Helpers.is(key, 'boolean'),
                 check = val !== Helpers._null ? val : (isbool ? key : true);
             return !isbool || val !== Helpers._null ? { [key] : { '$exists': check } } : { '$exists': check };
         },
-        fn: (val, bool, keys = Helpers._null, entry) => {
+        fn: function (val, bool, keys = Helpers._null, entry) {
             try {
                return bool === (Helpers.dotNotation(keys, entry[1]) !== undefined);
              } catch (error) {
@@ -21,7 +21,7 @@ module.exports = {
         }
     },
     '$type': {
-        fn: (val, type) => {
+        fn: function (val, type) {
             return Helpers.is(val, type);
         }
     }
